+++ lib/widget_extensions.dart
@@ -3,3 +3,4 @@ library widget_extensions;
 export 'extensions/common.dart';
 export 'extensions/container.dart';
 export 'extensions/text.dart';
+export 'extensions/others.dart';