+++ CHANGELOG.md
@@ -1,5 +1,18 @@
+## [0.0.2] : 08-12-2019
+
+* Added Row Extension Method
+* Added Column Extension Method
+* Added Stack Extension Method
+* Added Centre Extension Method
+* Added Wrap Extension Method
+* Added Opacity Extension Method
+* Added Align Extension Method
+* Added Padding Extension Method
+
 ## [0.0.1] : 02-12-2019
 
 * Initial Release
 * Added Container Extension Method
 * Added Text Extension Method
+* Added TextSpan Extension Method
+* Added TextStyle Extension Method